Implement basic settings in JoySensorConfigDialog
Show raw and processed sensor values with correct units in indicators.
Provide defaults for sensor parameters and allow changing them with
sliders and spin boxes. Use float values for parameters because sensor
values are floats. Finally, implement XML load/store for all new
values.

Presets and sensor preview widget will be implemented in a later commit.

+++ src/gui/joysensoreditdialog.h
@@ -51,7 +51,7 @@ class JoySensorEditDialog : public QDialog
 
     void refreshSensorStats(float x, float y, float z);
 
-    void checkMaxZone(int value);
+    void checkMaxZone(float value);
     void openMouseSettingsDialog();
     void enableMouseSettingButton();
     void updateWindowTitleSensorName();
